turned leaving into a int to distinguish between leaving angry and happily
in order to have different colors for when a customer is leaving according to whether they are angry or happy leaving had to take on three values not 2. -1 indicates not leaving, 0 indicates leaving angrily, 1 indicates leaving happily. leave() now takes in an int for the newLeaving value to distinguish between calling leave() in a happy or angry scenario